Dockerize one (first) microservice: Users
On how to create the JAR file for the Users microservice project from console.
Create a Dockerfile that copies the JAR into the image.
Set an entrypoint in the Dockerfile that run the JAR with the command 'java'.
References to 'localhost' in the project code must be changed to 'host.docker.internal'
(in application.properties and @FeignClient annotationes).